multi_agent_qa/requirements.txt
# Multi-Agent QA System Requirements

# Core dependencies
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.3.0

# Web frontend
flask>=2.3.0
flask-socketio>=5.3.0
eventlet>=0.33.0

# Data processing and utilities
numpy>=1.21.0
pandas>=1.3.0
pillow>=8.3.0
requests>=2.26.0

# Image processing and OCR (for screenshot analysis)
pytesseract>=0.3.8
opencv-python>=4.5.0

# Async and concurrency
aiohttp>=3.8.0
asyncio-throttle>=1.0.0

# Testing and validation
pytest>=6.2.0
pytest-asyncio>=0.18.0

# Logging and monitoring
loguru>=0.6.0
rich>=12.0.0

# Configuration and serialization
pydantic>=1.10.0
pyyaml>=6.0
toml>=0.10.0

# Optional dependencies for enhanced functionality
# Uncomment as needed

# For advanced ML/AI features
# torch>=1.12.0
# transformers>=4.20.0
# sentence-transformers>=2.2.0

# For computer vision enhancements
# mediapipe>=0.8.0
# ultralytics>=8.0.0

# For real Android device integration
# adb-shell>=0.4.0
# pure-python-adb>=0.3.0

# For web scraping and additional data sources
# beautifulsoup4>=4.11.0
# selenium>=4.0.0

# For advanced analytics and visualization
# matplotlib>=3.5.0
# seaborn>=0.11.0
# plotly>=5.0.0

# For database integration (optional)
# sqlalchemy>=1.4.0
# sqlite3  # Part of Python standard library

# Development dependencies
black>=22.0.0
flake8>=4.0.0
mypy>=0.950
pre-commit>=2.17.0

# Documentation
sphinx>=4.5.0
sphinx-rtd-theme>=1.0.0
